Refactor scripts

+++ manage.py
@@ -1,11 +1,14 @@
 #! /usr/bin/env python
-from flask.ext.script import Manager
+from flask.ext.script import Manager, prompt_bool
 from naverwebtoonfeeds import app
 
 manager = Manager(app)
 
-@manager.command
-def db_create_all():
+db_manager = Manager()
+manager.add_command('db', db_manager)
+
+@db_manager.command
+def create():
     """
     Create database tables. First check for the existence of each individual
     table, and if not found will issue the CREATE statements.
@@ -14,16 +17,18 @@ def db_create_all():
     from naverwebtoonfeeds import db
     db.create_all()
 
-@manager.command
-def db_drop_all():
+@db_manager.command
+def drop():
     """Drop all database tables."""
+    from naverwebtoonfeeds import db
     try:
-        db.drop_all()
+        if prompt_bool("Are you sure you want to lose all your data"):
+            db.drop_all()
     except:
         pass
 
-@manager.command
-def update():
+@db_manager.command
+def fill():
     """Update database by fetching changes from Naver Comics."""
     from naverwebtoonfeeds import cache
     from naverwebtoonfeeds.lib.updater import update_series_list
@@ -34,22 +39,11 @@ def update():
         cache.delete('feed_show_%d' % series_id)
     add_completed_series()
 
-@manager.command
-def add_completed_series():
-    """Add completed series."""
-    from naverwebtoonfeeds import cache, db
-    from naverwebtoonfeeds.models import Series
-    from naverwebtoonfeeds.lib.updater import __browser__, update_series
-    completed_series_ids = set(data['id'] for data in __browser__.get_completed_series())
-    existing_series_ids = set(row[0] for row in db.session.query(Series.id))
-    for series_id in completed_series_ids - existing_series_ids:
-        series = Series(id=series_id)
-        series.new_chapters_available = True
-        update_series(series)
-    cache.delete('feed_index')
+cache_manager = Manager()
+manager.add_command('cache', cache_manager)
 
-@manager.command
-def delete_cache(target='index'):
+@cache_manager.command
+def delete(target='index'):
     """
     Delete specified view cache. Call with the ID of the series to delete
     the view cache. If the target is not present or it is 'index', then the
@@ -67,6 +61,20 @@ def migrate(action):
     from flask.ext.evolution import Evolution
     evolution = Evolution(app)
     evolution.manager(action)
+
+@manager.command
+def addcompletedseries():
+    """Add completed series."""
+    from naverwebtoonfeeds import cache, db
+    from naverwebtoonfeeds.models import Series
+    from naverwebtoonfeeds.lib.updater import __browser__, update_series
+    completed_series_ids = set(data['id'] for data in __browser__.get_completed_series())
+    existing_series_ids = set(row[0] for row in db.session.query(Series.id))
+    for series_id in completed_series_ids - existing_series_ids:
+        series = Series(id=series_id)
+        series.new_chapters_available = True
+        update_series(series)
+    cache.delete('feed_index')
 
 if __name__ == '__main__':
     manager.run()
